+++ tools4j-collections/src/main/java/com/github/panpf/tools4j/collections/Grouping.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.panpf.tools4j.grouping;
+package com.github.panpf.tools4j.collections;
 
 import org.jetbrains.annotations.NotNull;