+++ tools4j-collections/src/main/java/com/github/panpf/tools4j/collections/Ref.java
@@ -1,4 +1,4 @@
-package com.github.panpf.tools4j.grouping;
+package com.github.panpf.tools4j.collections;
 
 import java.io.Serializable;